Update 3D) Best AM peak-to-midday transit travel time ratio
Previously, it was the max(wTrnW ratio) for the o/d TAZs in the o/d cities
Now, we just report the wTrnW ratio for each o/d TAZ with the lowest walk-transit-walk time in AM, and similarly for the midday. (These may be the same O/D pairs.)  Also report what those TAZs are.